% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{AddNewReference}
\alias{AddNewReference}
\title{AddNewReference
This function will create or increase the Exon counts data base.}
\usage{
AddNewReference(db, sbjsBamFiles, exonsBed = NULL, saveTo = NULL)
}
\arguments{
\item{db}{exon counts database}

\item{sbjsBamFiles}{path to wubjects}

\item{exonsBed}{it is required that the first three columns should be chromosome, start and end}

\item{save}{TRUE if you want to save the Database}
}
\description{
AddNewReference
This function will create or increase the Exon counts data base.
}
