% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Centromere}
\alias{Centromere}
\title{Centromere}
\format{
A data frame with 61 rows (neoantigens) and 10 variables (neoantigens info)
}
\usage{
data(Centromere)
}
\description{
Information regarding to Centromere position for hg19 and GRCh38
}
\details{
Is a list with two data.frames with three columns "chr", left , right (the last two identifies the position of the base partition. 
left correspond to the p. and right to q)
}
\keyword{datasets}
