% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{clean_field}
\alias{clean_field}
\title{Clean CNV fields}
\usage{
clean_field(x, integer = TRUE)
}
\description{
Clean CNV fields
}
\examples{
x <- c('snp=1', 'snp=2')
clean_field(x)

x <- c('snp=snp1', 'snp=snp2')
clean_field(x, FALSE)

}
