% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{CNVcall}
\alias{CNVcall}
\title{CNVcall  
This function execute the CNV call based on ExomeDepth. 
It assumes that the Subject is not present on the current Exon Count DB.
If the subjects is already in the DB, it will stop.}
\usage{
CNVcall(db, sbjsBamFiles, reference = c("auto", "all"))
}
\arguments{
\item{db}{exon counts database}

\item{sbjsBamFiles}{subject BAM files}

\item{reference}{the reference type}
}
\description{
CNVcall  
This function execute the CNV call based on ExomeDepth. 
It assumes that the Subject is not present on the current Exon Count DB.
If the subjects is already in the DB, it will stop.
}
