% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{CNVcallFromDB}
\alias{CNVcallFromDB}
\title{CNVcallFromDB  
This function execute the CNV call based on ExomeDepth. The subjects should be already in the Exon DB count database \code{\link{LoadDB}}}
\usage{
CNVcallFromDB(db, dbSubjects, reference = c("auto","all"))
It assumes that the Subject is not present on the current Exon Count DB.
If the subjects is already in the DB, it will stop.
}
\arguments{
\item{db}{exon counts database}

\item{dbSubjects}{the subjects to test}

\item{reference}{the reference type}
}
\description{
CNVcallFromDB  
This function execute the CNV call based on ExomeDepth. The subjects should be already in the Exon DB count database \code{\link{LoadDB}}
}
\details{
The subjects should be already in the DB by calling \code{\link{AddNewReference}}. Then the test subject should 
be identified by it colnames(db$DB).
if moretha one subject is referred, then all the subjects in the DB, except the one tested, 
will be used to build the reference according to the "reference" type used.
if reference == auto , then the reference will be chosen by correlation
if reference == all , all the remainders subjects will be summed up to build the reference
}
\examples{
\dontrun{
DB <- LoadDB("path/to/db/RDS")
DB <- AddNewReference(db=DB, 
sbjsBamFiles = "/..../PATIENTS/27040/27040_1.fastq.gz.subread.BAM",save=TRUE)
CNVs <- CNVcallFromDB(DB, "27040")
}
}
