% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{CNVReport}
\alias{CNVReport}
\title{CNVReport}
\usage{
CNVReport(cnv)
}
\arguments{
\item{cnv}{an ExomeDepth CNV call object \code{\link[ExomeDepth]{CallCNVs}}}

\item{sbjPath}{the subject path directory}
}
\value{
It creates an Excel file in the subject directory. The file is named "subject_CNVs.xlsx"
}
\description{
CNVReport
}
\examples{
\dontrun{
DB <- LoadDB("path/to/db/RDS")
DB <- AddNewReference(db=DB, 
sbjsBamFiles = "/..../PATIENTS/27040/27040_1.fastq.gz.subread.BAM",save=TRUE)
CNVs <- CNVcallFromDB(DB, "27040")
CNVReport(CNVs[[1]])
}
}
