% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryfuncions.cnvr.R
\name{.FormatGenecodeColumn}
\alias{.FormatGenecodeColumn}
\title{.FormatGenedodeColumn
internal function to format the output of CNVcalls}
\usage{
.FormatGenecodeColumn(genecode)
}
\arguments{
\item{genecode}{the column of the CNVcall object}
}
\description{
.FormatGenedodeColumn
internal function to format the output of CNVcalls
}
