% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{.FormatReport}
\alias{.FormatReport}
\title{.FormatReport
Internal function to order and format the final Excel report file.}
\usage{
.FormatReport(.data)
}
\arguments{
\item{.data}{an ExomeDepth CNV call object \code{\link[ExomeDepth]{CallCNVs}}}
}
\value{
an ordered and formatted data.frame
the genecode is formatted to diminish space and to facilitate its read.
The CNVs are ordered according to the bp size width
}
\description{
.FormatReport
Internal function to order and format the final Excel report file.
}
