% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{.Validate}
\alias{.Validate}
\title{.Validate
it validate the existence of the data base as well as its correct class
this is an internal function, not for users}
\usage{
.Validate(db)
}
\arguments{
\item{db}{an CNVRdb object}
}
\value{
it returns FALSE if all the elements are NULL, TRUE otherwise or stop if the db is not CNVRdb
}
\description{
.Validate
it validate the existence of the data base as well as its correct class
this is an internal function, not for users
}
