% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneTargetCNVs}
\alias{GeneTargetCNVs}
\title{GeneTargetCNVs}
\format{
A data frame
}
\usage{
data(GeneTargetCNVs)
}
\description{
Information about interesting genes to be pointed into the plots
}
\keyword{datasets}
