% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{get_genemodel}
\alias{get_genemodel}
\title{Get Gene Model}
\usage{
get_genemodel(txdb, org, symbol)
}
\arguments{
\item{txdb}{A TxDb object}

\item{org}{An OrgDb object}

\item{symbol}{A character gene symbol.}
}
\description{
Extracts the coding regions (exons), and the gaps between them (introns) from
a TxDb object.
}
\examples{
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
org <- org.Hs.eg.db::org.Hs.eg.db
gr <- get_genemodel(txdb, org, c('OR4C11', 'SLX4IP'))

}
