% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{get_overlap}
\alias{get_overlap}
\title{Get Overlap}
\usage{
get_overlap(cnv, signal, flank = 1000)
}
\arguments{
\item{cnv}{A GRanges object}

\item{signal}{A GRanges object}
}
\description{
Expands the `cnv` GRanges object with amount `flank` on both sides. Then the
overlap between the two objects is extracted.
}
\examples{
# signal
fl <- system.file('extdata/offspring.txt', package = 'cnvr')
col_names <- c('name', 'chr', 'pos', 'gt', 'lrr', 'baf')
signal <- read_signal(fl, col_names)

# cnv
fl <- system.file('extdata/family.gene.cnv', package = 'cnvr')
col_names <- c('region', 'numsnp', 'length', 'cn', 'sample', 'startsnp', 'endsnp', 'conf', 'gene', 'distance')
cnv <- read_cnv(fl, col_names)

get_overlap(cnv, signal)

}
