% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{GetCNVsAnnotation}
\alias{GetCNVsAnnotation}
\title{GetCNVsAnnotation}
\usage{
GetCNVsAnnotation(chr, x, countThreshold = 10)
}
\arguments{
\item{chr}{(chromosome) any of chr1..chr22}

\item{x}{a CNV object from ExonDepth}

\item{countThreshold}{(default = 10). The minimum expected exon count}
}
\value{
an annotation data frame for the specified chromosome to be used for plotting
}
\description{
GetCNVsAnnotation
}
