% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{LoadDB}
\alias{LoadDB}
\title{LoadDB}
\usage{
LoadDB(path)
}
\arguments{
\item{path}{a path direcotry to the DB}
}
\value{
an object of class CNVRdb, derived from a list with the following slots
DB : a data.table with chromosome star end name subject1 subject2 .....
path : path
GenomeDBversion
}
\description{
LoadDB
}
