% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_bands}
\alias{plot_bands}
\title{Plot cytobands}
\usage{
plot_bands(gr, shade = TRUE, names = TRUE)
}
\arguments{
\item{gr}{A GRanges object}

\item{shade}{A Logical, default TRUE. Shade bands}

\item{names}{A Logical, default TRUE. Plot band names}
}
\value{
A plot
}
\description{
Plot cytobands
}
\examples{
x <- 1:10
set.seed(1234)
y <- rnorm(10)
b <- tibble::tibble(
  chrom = rep('chr1', 3),
  start = c(1, 4, 8),
  end = c(4, 8, 10),
  band = c('p1', 'p2', 'p3'),
  stain = c('gneg', 'gpos', 'gneg')
)
plot(x, y, type = 'n')
plot_bands(b)
points(x, y)

}
