% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot Gene Model}
\usage{
plot_gene(gr, xlim, ...)
}
\arguments{
\item{gr}{A GRanges object}

\item{xlim}{A numeric vector of length 2.}
}
\description{
Plot the exons and introns as rectangles and lines.
}
\examples{
d <- data.frame(
  seqnames = rep('chr1', 7),
  start = c(10, 40, 60, 30, 50, 05, 80),
  end   = c(30, 50, 80, 40, 60, 10, 90),
  feature  = rep(c('coding_region', 'intron', "5' utr", "3' utr"), c(3,2,1,1)),
  gene = rep('gene1', 7)
)

gr1 <- GenomicRanges::makeGRangesFromDataFrame(d, keep.extra.columns = TRUE)
gr2 <- GenomicRanges::shift(gr1, 10)
gr2$gene <- 'gene2'

gr <- c(gr1, gr2)

plot_gene(gr, c(0, 100))

}
