% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot Signal}
\usage{
plot_signal(
  signal,
  cnv,
  flank = 2e+05,
  type = "LRR",
  gene_model = NULL,
  bands = NULL,
  ...
)
}
\arguments{
\item{signal}{A `GRanges` object.}

\item{cnv}{A `GRanges` object.}

\item{type}{Default `'LRR'`. Other accepted signals are `'BAF'`}

\item{gene_model}{A `GRanges` object.}

\item{bands}{A `GRanges` object.}

\item{...}{Other arguemnts passed to `plot`}
}
\description{
Plots the position in the `signal` `GRanges` object at the x-axis and a
metadata column at the y-axis: lrr or baf. When `plot_gene = TRUE` a `GRanges`
object with the gene model when provided.
}
\examples{
# get gene model
gene <- 'SLX4IP'
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
org <- org.Hs.eg.db::org.Hs.eg.db
gene_model <- get_genemodel(txdb, org, gene)

# signal
fl <- system.file('extdata/offspring.txt', package = 'cnvr')
col_names <- c('name', 'chr', 'pos', 'gt', 'lrr', 'baf')
signal <- read_signal(fl, col_names)

# cnv
fl <- system.file('extdata/family.gene.cnv', package = 'cnvr')
col_names <- c('region', 'numsnp', 'length', 'cn', 'sample', 'startsnp', 'endsnp', 'conf', 'gene', 'distance')
cnv <- read_cnv(fl, col_names)

sample <- "data-raw/PennCNV/example/offspring.txt"
cn <- 1

cnv <- cnv[cnv$sample == sample & cnv$cn == cn & grepl(gene, cnv$gene)]
cnv$gene <- gene

flank <- 200000
# read cytobands file
fl <- system.file('extdata/hg38.cytoBand.txt', package = 'cnvr')
cytobands <- read_cytobands(fl)
cytobands <- IRanges::subsetByOverlaps(cytobands, cnv)

# plot
plot_signal(signal, cnv, flank, type = 'LRR', gene_model, bands = cytobands)
plot_signal(signal, cnv, flank, type = 'BAF', gene_model, bands = cytobands)

}
