% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_cnv}
\alias{read_cnv}
\title{Read CNV file}
\usage{
read_cnv(
  file,
  col_names = FALSE,
  tidy = TRUE,
  as.granges = TRUE,
  numeric_columns = c("numsnp", "length", "cn", "conf"),
  character_columns = c("startsnp", "endsnp"),
  style = "UCSC",
  ...
)
}
\arguments{
\item{file}{A tab separated file}

\item{col_names}{A character vector with length equal to the `file` columns.}

\item{tidy}{Default TRUE.}

\item{as.granges}{Default TRUE.}

\item{numeric_columns}{A character vector. Columns to clean as numeric}

\item{character_columns}{A character vector. Columns to clean as character}

\item{style}{Default 'UCSC'}

\item{...}{Other arguments passed to `read_tsv`}
}
\description{
Expects tab delimited files similar to PennCNV output. Columns names can be
edited using the `col_names` argument.
}
\examples{
# locate cnv file
fl <- system.file('extdata/offspring.cnv', package = 'cnvr')
col_names <- c('region', 'numsnp', 'length', 'cn', 'sample', 'startsnp', 'endsnp', 'conf')

# read cnv file
read_cnv(fl, col_names)

}
