% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_cytobands}
\alias{read_cytobands}
\title{Read cytoBands file}
\usage{
read_cytobands(file, as.granges = TRUE, style = "UCSC", ...)
}
\arguments{
\item{file}{A tab separated file}

\item{as.granges}{Default TRUE.}

\item{style}{Default 'UCSC'}

\item{...}{Other arguments passed to `read_tsv`}
}
\description{
Expects a tab separate file with at 5 columns: chrom, start, end, band, stain.
}
\examples{
# locate cytobands file
fl <- system.file('extdata/hg38.cytoBand.txt', package = 'cnvr')

# read cytobands file
read_cytobands(fl)

}
