% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_pfb}
\alias{read_pfb}
\title{Read PFB file}
\usage{
read_pfb(file, col_names = NULL, as.granges = FALSE, style = "UCSC", ...)
}
\arguments{
\item{file}{A tab separated file}

\item{col_names}{A character vector with length equal to the `file` columns.}

\item{as.granges}{Default FALSE.}

\item{style}{Default 'UCSC'}

\item{...}{Other arguments passed to `read_tsv`}
}
\description{
Expects a tab separate file with at 4 columns: name, chr, pos, and pfb.
Column names can be changed using the `col_names` argument.
}
\examples{
# locate pfb file
fl <- system.file('extdata/example.pfb', package = 'cnvr')

# read pfb file
col_names <- c('name', 'chr', 'pos', 'pfb')
read_pfb(fl, col_names)
read_pfb(fl, col_names, as.granges = TRUE)

}
