% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_signal}
\alias{read_signal}
\title{Read Signal}
\usage{
read_signal(
  file,
  col_names = NULL,
  pfb = NULL,
  as.granges = TRUE,
  style = "UCSC",
  ...
)
}
\arguments{
\item{file}{A tab separated file}

\item{col_names}{A character vector with length equal to the `file` columns.}

\item{as.granges}{Default TRUE.}

\item{style}{Default 'UCSC'}

\item{...}{Other arguments passed to `read_tsv`}
}
\description{
Expects a tab separate file with at 5 columns: name, chr, pos, lrr, and baf.
Column names can be changed using the `col_names` argument.
}
\examples{
# locate signal file
fl <- system.file('extdata/offspring.txt', package = 'cnvr')

# read signal file
col_names <- c('name', 'chr', 'pos', 'gt', 'lrr', 'baf')
read_signal(fl, col_names)

# no pos
fl <- system.file('extdata/example.pfb', package = 'cnvr')
col_names <- c('name', 'chr', 'pos', 'pfb')
pfb <- read_pfb(fl, col_names)

# read signal file, no pos
fl <- system.file('extdata/offspring.nopos.txt', package = 'cnvr')
col_names <- c('name', 'gt', 'lrr', 'baf')
read_signal(fl, col_names, pfb)

}
