% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{AddNewReference}
\alias{AddNewReference}
\title{AddNewReference
This function will create or increase the Exon counts data base.}
\usage{
AddNewReference(db, sbjsBamFiles, exonsBed = NULL, saveTo = NULL)
}
\arguments{
\item{db}{exon counts database}

\item{sbjsBamFiles}{path to subjects}

\item{exonsBed}{it is required that the first three columns should be chromosome, start and end}

\item{saveTo}{(character) default NULL. The path to save the database. It is only valid if the database do not exists}
}
\description{
AddNewReference
This function will create or increase the Exon counts data base.
}
