% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{CNVcall}
\alias{CNVcall}
\title{CNVcall}
\usage{
CNVcall(db, sbjsBamFiles, reference = c("auto", "all"))
}
\arguments{
\item{db}{exon counts database}

\item{sbjsBamFiles}{subject BAM files}

\item{reference}{the reference type}
}
\description{
CNVcall
}
