% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVR.utils.R
\name{LoadDB}
\alias{LoadDB}
\title{LoadDB}
\usage{
LoadDB(path)
}
\arguments{
\item{path}{a path directory to the DB}
}
\value{
an object of class CNVRdb, derived from a list with the following slots
It should hold the following slots:
DB : (data.matrix) the exon counts data matrix of the reference subjects (ExS)
exons : (data.frame) annotation information chromosome, start, end , strand , ...
path : (character) full path where the library is saved
GenomeDBversion : the genome DB version (see \code{\link[GenomeDB]{GetGenome}}) (all the samples should be processed by the same GenomeDBversion)
Aligner : the used aligner (all the samples should be processed by the same Aligner)
}
\description{
LoadDB
}
