% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(cnv, top_samples = NULL, top_genes = NULL, ...)
}
\arguments{
\item{cnv}{A GRanges object such as the one created using `read_cnv`}

\item{top_samples}{An integer. The number of top samples to plot}

\item{top_genes}{An integer. The number of top genes to plot}

\item{...}{Other arguments passed to `Heatmap`}
}
\description{
Expects an annotated PennCNV output. The `gene` and `sample` columns are used
to create a matrix with the number/type of CNVs. The matrix is plotted as a
heatmap with the number of CNVs in each gene on the right as a bar plot and
the number of affected genes in each sample on the top.
}
\examples{
# locate and read cnv file
fl <- system.file('extdata/family.gene.cnv', package = 'cnvr')
col_names <- c('region', 'numsnp', 'length', 'cn', 'sample', 'startsnp', 'endsnp', 'conf', 'gene', 'distance')
cnv <- read_cnv(fl, col_names)
cnv <- cnv[cnv$gene != 'NOT_FOUND']

# plot heatmap
plot_heatmap(cnv)

}
