% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnv.R
\name{plot_signal}
\alias{plot_signal}
\title{Plot Signal}
\usage{
plot_signal(gr, type = "LRR", gene_model = NULL, plot_gene = TRUE, ...)
}
\arguments{
\item{gr}{A `GRanges` object.}

\item{type}{Default `'LRR'`. Other accepted signals are `'BAF'`}

\item{gene_model}{A `GRanges` object.}

\item{plot_gene}{Default TRUE.}

\item{...}{Other arguemnts passed to `plot`}
}
\description{
Plots the position in the `gr` object at the x-axis and a metadata column at
the y-axis: lrr or baf. When `plot_gene = TRUE` a GRanges object with the
gene model is expected and plotted using `plot_gene` underneath the signal
panel.
}
\examples{
# get gene model
txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene::TxDb.Hsapiens.UCSC.hg38.knownGene
org <- org.Hs.eg.db::org.Hs.eg.db
gene_model <- get_genemodel(txdb, org, 'OR4C11')

# signal
fl <- system.file('extdata/offspring.txt', package = 'cnvr')
col_names <- c('name', 'chr', 'pos', 'gt', 'lrr', 'baf')
signal <- read_signal(fl, col_names)

# cnv
fl <- system.file('extdata/family.gene.cnv', package = 'cnvr')
col_names <- c('region', 'numsnp', 'length', 'cn', 'sample', 'startsnp', 'endsnp', 'conf', 'gene', 'exon')
cnv <- read_cnv(fl, col_names)
cnv <- cnv[grepl('OR4C11', cnv$gene) & cnv$cn == 0]

# get the overlap
ol <- get_overlap(cnv, signal, flank = 5000)

plot_signal(ol, type = 'LRR', gene_model)
plot_signal(ol, type = 'BAF', gene_model)

}
